refactoring fetch-transport.ts out of logging folder
