+++ packages/crashlytics/src/logging/fetch-transport.test.ts
@@ -21,7 +21,7 @@
 import * as sinon from 'sinon';
 import * as assert from 'assert';
 import { DynamicHeaderProvider } from '../types';
-import { FetchTransport } from './fetch-transport';
+import { FetchTransport } from '../fetch-transport';
 import {
   ExportResponseRetryable,
   ExportResponseFailure,
